20191111|django form 합치기

+++ 05_django/04_django_form/articles/views.py
@@ -1,15 +1,23 @@
-from django.views.decorators.http import require_POST
-from django.shortcuts import render, redirect, get_object_or_404
+import hashlib
 from IPython import embed
+from django.views.decorators.http import require_POST
+from django.contrib.auth.decorators import login_required
+from django.shortcuts import render, redirect, get_object_or_404
 from .models import Article, Comment
 from .forms import ArticleForm, CommentForm
 
 # Create your views here.
 def index(request):
+  # if request.user.is_authenticated:
+  #   gravatar_url = hashlib.md5(request.user.email.encode('utf-8').lower().strip()).hexdigest()
+  # else:
+  #   gravatar_url = None
+
   articles = Article.objects.all()[::-1]
   context = {'articles':articles}
   return render(request,'articles/index.html',context)
 
+@login_required
 def create(request):
   #사용자로부터 데이터를 받아서 DB에 저장하는 함수
   if request.method == 'POST':
@@ -17,7 +25,7 @@ def create(request):
     # 폼 인스턴스를 생성하고, 전달받은 데이터를 채운다.
     # 인스턴스에 데이터를 채워서, 유효성 검증을 진행한다.
     form = ArticleForm(request.POST)
-    #embed()
+    # embed()
     if form.is_valid():
       # cleaned_data를 통해 딕셔너리 안 데이터를 검증한다.
       # title = form.cleaned_data.get('title')
@@ -49,11 +57,12 @@ def detail(request,article_pk):
 
 @require_POST
 def delete(request, article_pk):
-  article = get_object_or_404(Article, pk=article_pk)
-  article.delete()
+  if request.user.is_authenticated:
+    article = get_object_or_404(Article, pk=article_pk)
+    article.delete()
   return redirect('articles:index')
   
-
+@login_required
 def update(request, article_pk):
   article = get_object_or_404(Article, pk=article_pk)
   if request.method == 'POST':
@@ -76,19 +85,21 @@ def update(request, article_pk):
 @require_POST
 def comment_create(request, article_pk):
   article = get_object_or_404(Article, pk=article_pk)
-  comment_form = CommentForm(request.POST)
-  if comment_form.is_valid():
-    # save 메서드 -> 선택 인자 : (기본값) commit=True
-    # commit=False : DB에 바로 저장되는 것을 막아준다.
-    comment = comment_form.save(commit=False)
-    comment.article = article
-    comment.save()
-  return redirect('articles:detail', article.pk)
+  if request.user.is_authenticated:
+    comment_form = CommentForm(request.POST)
+    if comment_form.is_valid():
+      # save 메서드 -> 선택 인자 : (기본값) commit=True
+      # commit=False : DB에 바로 저장되는 것을 막아준다.
+      comment = comment_form.save(commit=False)
+      comment.article = article
+      comment.save()
+  return redirect(request.GET.get('next') or 'articles:detail', article.pk)
 
 
 # 댓글 삭제 뷰 함수
 @require_POST
 def comment_delete(request, article_pk, comment_pk):
-  comment = get_object_or_404(Comment, pk=comment_pk)
-  comment.delete()
+  if request.user.is_authenticated:
+    comment = get_object_or_404(Comment, pk=comment_pk)
+    comment.delete()
   return redirect('articles:detail', article_pk)
